:recycle: rename package server to minecraftserver
