+++ server_handler/internal/minecraftserver/server.go
@@ -1,4 +1,4 @@
-package server
+package minecraftserver
 
 import (
 	"fmt"
